+++ metric/categorical_macro_precision.go
@@ -6,7 +6,7 @@ import "github.com/itsmontoya/neuralnetwork/matrix"
 type CategoricalMacroPrecision struct{}
 
 // Value returns macro-averaged precision across classes.
-func (c CategoricalMacroPrecision) Value(predictions, targets *matrix.Matrix) (value float64, err error) {
+func (c CategoricalMacroPrecision) Value(predictions, targets *matrix.Matrix) (value float32, err error) {
 	var confusionMatrix *ConfusionMatrix
 
 	if confusionMatrix, err = NewCategoricalConfusionMatrix(predictions, targets); err != nil {
