+++ metric/metric.go
@@ -6,5 +6,5 @@ import "github.com/itsmontoya/neuralnetwork/matrix"
 // Metric evaluates predictions against targets for reporting.
 type Metric interface {
 	// Value computes a reporting metric from predictions and targets.
-	Value(predictions, targets *matrix.Matrix) (value float64, err error)
+	Value(predictions, targets *matrix.Matrix) (value float32, err error)
 }
