+++ model/accuracy_func.go
@@ -3,4 +3,4 @@ package model
 import "github.com/itsmontoya/neuralnetwork/matrix"
 
 // AccuracyFunc computes an accuracy value from predictions and targets.
-type AccuracyFunc func(predictions, targets *matrix.Matrix) (accuracy float64, err error)
+type AccuracyFunc func(predictions, targets *matrix.Matrix) (accuracy float32, err error)
